// src/types/interface.ts
export interface IMasul {
  id: number;
  label: string;
  lavozim: string;
  unvon: string;
  fish: string;
}

export interface INavbatchi {
  id: number;
  label: string;
  unvon: string;
  fish: string;
}

export interface ITergov {
  id: number;
  tb: string;
  jqb: string;
  ekb: string;
  vaqt: string;
  xaydovchi: string;
}

export interface INojXodim {
  id: number;
  fish: string;
  tel: string;
}

export interface IPPX1Item {
  id: number;
  iib: string;
  tel: string;
  hudud: string;
  xodim1: string;
  xodim2: string;
  mashina: string;
  raqam: string;
}

export interface IPPX2Item {
  id: number;
  chr: string;
  hudud: string;
  xodim1: string;
  xodim2: string;
  izoh: string;
}

export interface IXPBItem {
  id: number;
  iib: string;
  tel: string;
  hudud: string;
  xodim1: string;
  xodim2: string;
  mashina: string;
  raqam: string;
}